requirements.txt
anyio==4.10.0
APScheduler==3.11.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
numpy==2.3.3
pandas==2.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-telegram-bot==22.4
pytz==2025.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
telegram==0.0.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
